Moved FakeCrm work into module
Allows it to be conditionally loaded
