+++ config/environments/development.rb
@@ -85,4 +85,6 @@ Rails.application.configure do
   config.x.oidc_client_id = 'schoolexperience'
   config.x.oidc_client_secret = Rails.application.credentials.dig(:dfe_pp_signin_secret)
   config.x.oidc_host = 'pp-oidc.signin.education.gov.uk'
+
+  config.x.fake_crm = true
 end
