+++ config/environments/production.rb
@@ -129,4 +129,6 @@ Rails.application.configure do
     end
     config.x.oidc_host = ENV.fetch('DFE_SIGNIN_HOST') { 'pp-oidc.signin.education.gov.uk' }
   end
+
+  config.x.fake_crm = true
 end
